Fixing bug in log file purge logic (#393)

+++ src/WebJobs.Script.WebHost/SecretManager.cs
@@ -7,6 +7,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
+using Microsoft.Azure.WebJobs.Host;
 using Newtonsoft.Json;
 
 namespace Microsoft.Azure.WebJobs.Script.WebHost
@@ -107,42 +108,56 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost
         /// to a function.
         /// </summary>
         /// <param name="rootScriptPath">The root function directory.</param>
-        public void PurgeOldFiles(string rootScriptPath)
+        /// <param name="traceWriter">The TraceWriter to log to.</param>
+        public void PurgeOldFiles(string rootScriptPath, TraceWriter traceWriter)
         {
-            if (!Directory.Exists(rootScriptPath))
+            try
             {
-                return;
+                if (!Directory.Exists(rootScriptPath))
+                {
+                    return;
+                }
+
+                // Create a lookup of all potential functions (whether they're valid or not)
+                // It is important that we determine functions based on the presence of a folder,
+                // not whether we've identified a valid function from that folder. This ensures
+                // that we don't delete logs/secrets for functions that transition into/out of
+                // invalid unparsable states.
+                var functionLookup = Directory.EnumerateDirectories(rootScriptPath).ToLookup(p => Path.GetFileName(p), StringComparer.OrdinalIgnoreCase);
+
+                var secretsDirectory = new DirectoryInfo(_secretsPath);
+                if (!Directory.Exists(_secretsPath))
+                {
+                    return;
+                }
+
+                foreach (var secretFile in secretsDirectory.GetFiles("*.json"))
+                {
+                    if (string.Compare(secretFile.Name, ScriptConstants.HostMetadataFileName, StringComparison.OrdinalIgnoreCase) == 0)
+                    {
+                        // the secrets directory contains the host secrets file in addition
+                        // to function secret files
+                        continue;
+                    }
+
+                    string fileName = Path.GetFileNameWithoutExtension(secretFile.Name);
+                    if (!functionLookup.Contains(fileName))
+                    {
+                        try
+                        {
+                            secretFile.Delete();
+                        }
+                        catch
+                        {
+                            // Purge is best effort
+                        }
+                    }
+                }
             }
-
-            // Create a lookup of all potential functions (whether they're valid or not)
-            // It is important that we determine functions based on the presence of a folder,
-            // not whether we've identified a valid function from that folder. This ensures
-            // that we don't delete logs/secrets for functions that transition into/out of
-            // invalid unparsable states.
-            var functionLookup = Directory.EnumerateDirectories(rootScriptPath).ToLookup(p => Path.GetFileName(p), StringComparer.OrdinalIgnoreCase);
-
-            var secretsDirectory = new DirectoryInfo(_secretsPath);
-            foreach (var secretFile in secretsDirectory.GetFiles("*.json"))
+            catch (Exception ex)
             {
-                if (string.Compare(secretFile.Name, ScriptConstants.HostMetadataFileName, StringComparison.OrdinalIgnoreCase) == 0)
-                {
-                    // the secrets directory contains the host secrets file in addition
-                    // to function secret files
-                    continue;
-                }
-
-                string fileName = Path.GetFileNameWithoutExtension(secretFile.Name);
-                if (!functionLookup.Contains(fileName))
-                {
-                    try
-                    {
-                        secretFile.Delete();
-                    }
-                    catch
-                    {
-                        // Purge is best effort
-                    }
-                }
+                // Purge is best effort
+                traceWriter.Error("An error occurred while purging secret files", ex);
             }
         }
